Enable [[trivial_abi]] for std::unique_ptr
One of the ways in that unique_ptr is not actually zero-cost is that because it
has a nontrivial destructor, the Itanium ABI disallows passing it in a
register. Clang's [[trivial_abi]] attribute makes the compiler ignore what the
ABI says and pass it as if it was a bare pointer, which results in smaller and
faster code. Because std::unique_ptr is a standard library type, we can't mark
it directly, but libc++ will mark it if the appropriate define is set.

Passing a unique_ptr between code built with this enabled and code built with
it disabled doesn't work, so realm-cpp can't use this.  No other SDKs have a
c++ API and so it should be fine everywhere else, so it is on by default.