Remove ArrayWithFind's callback parameter
This was down to being used in one place, and that one place wasn't actually
using it to avoid a virtual call per row. In addition, how it was being used (a
lambda inside of a templated function) resulted in a very large number of
instantiations of the ArrayWithFind functions rather than the expected two.

find_action_pattern() has been unused for several years,